Use the default NodeOutputBuilder::build method for NodeOutput.

+++ server/src/http_objects.rs
@@ -758,7 +758,7 @@ pub struct Tasks {
 pub struct FnOutput {
     pub id: String,
     pub compute_fn: String,
-    pub created_at: u64,
+    pub created_at: u128,
 }
 
 #[derive(Debug, Serialize, Deserialize, ToSchema, Clone)]
